Rename the whatwedo.js into services.js
